pyproject.toml
[project]
name = "synta-takehome"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.57.0",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
]
